+++ internal/extproc/translator/gemini_helper.go
@@ -11,10 +11,7 @@ import (
 	"mime"
 	"net/url"
 	"path"
-	"strconv"
 
-	"github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
-	"github.com/envoyproxy/go-control-plane/envoy/service/ext_proc/v3"
 	"github.com/google/uuid"
 	"google.golang.org/genai"
 
@@ -105,17 +102,6 @@ func openAIMessagesToGeminiContents(messages []openai.ChatCompletionMessageParam
 		gcpContents = append(gcpContents, genai.Content{Role: genai.RoleUser, Parts: gcpParts})
 	}
 	return gcpContents, systemInstruction, nil
-}
-
-// systemMsgToDeveloperMsg converts OpenAI system message to developer message.
-// Since systemMsg is deprecated, this function is provided to maintain backward compatibility.
-func systemMsgToDeveloperMsg(msg openai.ChatCompletionSystemMessageParam) openai.ChatCompletionDeveloperMessageParam {
-	// Convert OpenAI system message to developer message.
-	return openai.ChatCompletionDeveloperMessageParam{
-		Name:    msg.Name,
-		Role:    openai.ChatMessageRoleDeveloper,
-		Content: msg.Content,
-	}
 }
 
 // developerMsgToGeminiParts converts OpenAI developer message to Gemini Content.
@@ -480,46 +466,4 @@ func geminiLogprobsToOpenAILogprobs(logprobsResult genai.LogprobsResult) openai.
 func buildGCPModelPathSuffix(publisher, model, gcpMethod string) string {
 	pathSuffix := fmt.Sprintf("publishers/%s/models/%s:%s", publisher, model, gcpMethod)
 	return pathSuffix
-}
-
-// buildGCPRequestMutations creates header and body mutations for GCP requests
-// It sets the ":path" header, the "content-length" header and the request body.
-func buildGCPRequestMutations(path string, reqBody []byte) (*ext_procv3.HeaderMutation, *ext_procv3.BodyMutation) {
-	var bodyMutation *ext_procv3.BodyMutation
-	var headerMutation *ext_procv3.HeaderMutation
-
-	// Create header mutation.
-	if len(path) != 0 {
-		headerMutation = &ext_procv3.HeaderMutation{
-			SetHeaders: []*corev3.HeaderValueOption{
-				{
-					Header: &corev3.HeaderValue{
-						Key:      ":path",
-						RawValue: []byte(path),
-					},
-				},
-			},
-		}
-	}
-
-	// If the request body is not empty, we set the content-length header and create a body mutation.
-	if len(reqBody) != 0 {
-		if headerMutation == nil {
-			headerMutation = &ext_procv3.HeaderMutation{}
-		}
-		// Set the "content-length" header.
-		headerMutation.SetHeaders = append(headerMutation.SetHeaders, &corev3.HeaderValueOption{
-			Header: &corev3.HeaderValue{
-				Key:      HTTPHeaderKeyContentLength,
-				RawValue: []byte(strconv.Itoa(len(reqBody))),
-			},
-		})
-
-		// Create body mutation.
-		bodyMutation = &ext_procv3.BodyMutation{
-			Mutation: &ext_procv3.BodyMutation_Body{Body: reqBody},
-		}
-	}
-
-	return headerMutation, bodyMutation
 }
